Refactor: seperate StackInput as an independent component
